fix(ResultGroup): Overflow when get profits with first result
- 1등의 상금으로 인한 오버플로우 수정

+++ src/main/java/lotto/domain/ResultGroup.java
@@ -26,7 +26,7 @@ public class ResultGroup {
     public double calculateProfits(Money money) {
         double totalEarning = 0;
         for (WinningResult result : WinningResult.winningResults()) {
-            totalEarning += howManyHave(result) * result.prize();
+            totalEarning += howManyHave(result) * (double) result.prize();
         }
         return totalEarning / money.value();
     }
